Update readme, formatting, dependencies
This commit cleans up the readme and also addresses any
formatting and linting warnings.

When running the app in Windows, some of the dependencies
were required to be added in seperately. This commit also adds
those dependencies for windows compatibility.

Also ensures the linux line endings are set as the standard.