Add github social provider login
This commit adds the feature of being able to login using your
github account using OAuth and the passport-github2 strategy.

While both the local strategy and github strategy use the same
mongo db collection, we serialize the sessions on different things.
In the case of the github social provider, we end up serializing
on the github id, (an inherent assumption here is that it is unique
for each user and not re-used).

TODO's mark areas where additional features / refactoring is needed.