Tighten up the session cookie requirements
This commit adds additional protection from cookie by explicity
adding cookie attributes that include Secure:true, HostOnly:true,
HttpOnly:true, the path restricted to "/" and sameSite:"Strict".

It should be noted that the defaults for express.js are actually
all of the above, except the default for secure is false. By
defining these explicitly we can be reminded what they are set to.