Ensure state parameter is checked (github)
The previous implementation of the social login for github
did not include the state parameter, which certainly
could open us up to CSRF attacks.

The following commit sends the state option to the passport
session middleware to create state to verify. Also, after
tightening down the cookies and adding the state parameter,
I needed to lighten the sameSite requirement to lax in order
to allow session to flow back from github when we do
social auth.