Add scope support to OAuth Flows
This commit adds basic scope support. Scopes are saved
in the original request with the authorization code and
then saved with the access token when the token endpoint
is called.

This commit also fixes a small typo for the cancel
flow where the server.decision was not picking up the
value because it was capitalized.

This commit also adds the get and post /userinfo endpoint
by mocking some data and using data in the request to
return to the user based upon the profile and/or
email scopes.

If the scope is not openid, then the endpoints will return
a 401.