Adds unique constraints to schema
While the username was already set to a unique constraint using
the passport-local-mongoose plugin, the githubId was not set as
a unique constraint. Since I made the assumption that the githubId
was unique, I should constrain it as such.

This also makes the username a unique constraint explicity, just
for clarity.